adafruit_minimqtt: Fix issue #86 with multiple poll() PINGREQs
If I set timeout to zero there is a problem as the poll() call will be
called multiple times while a ping request is outstanding and this
results in multiple ping requests being made. I see 3-5 at a time with
DEBUG logging on.

This is because self._timestamp is being reset after self.ping()

The solution seems to be pretty straightforward which is to reset
self._timestamp() before sending the ping()

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>