Create minimqtt_multipub_simpletest.py
Adds a simple demo for publishing multiple feeds. Designed to be used with the BME280 sensor but filled in with fake values so it should work as is. Most other demos are using the ESP32SPI library.  This one is more designed for the newer ESP32-S2 & S3 with no coprocessor.